Unify Logging, Improve Recovery, & Kill Helpers.cs
